+++ activesupport/lib/active_support/cache/file_store.rb
@@ -128,6 +128,8 @@ module ActiveSupport
               raise if File.exist?(key)
               false
             end
+          else
+            false
           end
         end