+++ actionpack/test/dispatch/system_testing/driver_test.rb
@@ -28,6 +28,14 @@ class DriverTest < ActiveSupport::TestCase
     assert_equal ({ url: "http://example.com/wd/hub" }), driver.instance_variable_get(:@options)
   end
 
+  test "initializing the driver with a headless chrome and custom path" do
+    original_driver_path = ::Selenium::WebDriver::Chrome::Service.driver_path
+    ::Selenium::WebDriver::Chrome::Service.driver_path = "bin/test"
+    ActionDispatch::SystemTesting::Driver.new(:selenium, using: :headless_chrome, screen_size: [1400, 1400])
+  ensure
+    ::Selenium::WebDriver::Chrome::Service.driver_path = original_driver_path
+  end
+
   test "initializing the driver with a headless firefox" do
     driver = ActionDispatch::SystemTesting::Driver.new(:selenium, using: :headless_firefox, screen_size: [1400, 1400], options: { url: "http://example.com/wd/hub" })
     assert_equal :selenium, driver.instance_variable_get(:@driver_type)
@@ -35,6 +43,14 @@ class DriverTest < ActiveSupport::TestCase
     assert_instance_of Selenium::WebDriver::Firefox::Options, driver.instance_variable_get(:@browser).options
     assert_equal [1400, 1400], driver.instance_variable_get(:@screen_size)
     assert_equal ({ url: "http://example.com/wd/hub" }), driver.instance_variable_get(:@options)
+  end
+
+  test "initializing the driver with a headless firefox and custom path" do
+    original_driver_path = ::Selenium::WebDriver::Firefox::Service.driver_path
+    ::Selenium::WebDriver::Firefox::Service.driver_path = "bin/test"
+    ActionDispatch::SystemTesting::Driver.new(:selenium, using: :headless_firefox, screen_size: [1400, 1400])
+  ensure
+    ::Selenium::WebDriver::Firefox::Service.driver_path = original_driver_path
   end
 
   test "initializing the driver with a cuprite" do