+++ activerecord/lib/active_record/relation/query_methods.rb
@@ -76,7 +76,7 @@ module ActiveRecord
         associations.each do |association|
           reflection = scope_association_reflection(association)
           @scope.joins!(association)
-          if @scope.table_name == reflection.table_name
+          if @scope.values.size > 1
             self.not(association => { reflection.association_primary_key => nil })
           else
             self.not(reflection.table_name => { reflection.association_primary_key => nil })
@@ -108,7 +108,7 @@ module ActiveRecord
         associations.each do |association|
           reflection = scope_association_reflection(association)
           @scope.left_outer_joins!(association)
-          if @scope.table_name == reflection.table_name
+          if @scope.values.size > 1
             @scope.where!(association => { reflection.association_primary_key => nil })
           else
             @scope.where!(reflection.table_name => { reflection.association_primary_key => nil })