+++ activesupport/lib/active_support/core_ext/date/deprecated_conversions.rb
@@ -15,7 +15,18 @@ class Date
         strftime(formatter)
       end
     elsif format == NOT_SET
-      to_default_s
+      if formatter = DATE_FORMATS[:default]
+        ActiveSupport::Deprecation.warn(
+          "Using a :default format for Date#to_s is deprecated. Please use Date#to_fs instead."
+        )
+        if formatter.respond_to?(:call)
+          formatter.call(self).to_s
+        else
+          strftime(formatter)
+        end
+      else
+        to_default_s
+      end
     else
       ActiveSupport::Deprecation.warn(
         "Date#to_s(#{format.inspect}) is deprecated. Please use Date#to_fs(#{format.inspect}) instead."