+++ actiontext/lib/action_text/gem_version.rb
@@ -9,7 +9,7 @@ module ActionText
   module VERSION
     MAJOR = 7
     MINOR = 0
-    TINY  = 6
+    TINY  = 7
     PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")